+++ NEWS.md
@@ -1,3 +1,9 @@
+# inlmisc 0.4.2.9000
+
+- In `GetTolColors` function: add color schemes;
+  change default scheme from `"bright"` to `"smooth rainbow"`;
+  add `start`, `end`, and `...` arguments.
+
 # inlmisc 0.4.2
 
 - In `GetTolColors` function, revise color schemes based on issue 3.0 of technical note.